Refacto login page component into two sub-component
