+++ src/pages/login/accountSelection/AccountSelection.css
@@ -1,0 +1,15 @@
+.action-container {
+  min-height: 125px;
+  justify-content: space-between;
+  padding: 10px;
+}
+
+.demo-link {
+  color: rgb(0, 255, 255);
+}
+
+@media screen and (min-width: 1025px) {
+  .action-container {
+    flex-direction: column-reverse;
+  }
+}
