+++ src/pages/login/authCode/AuthCode.css
@@ -1,0 +1,10 @@
+.action-container-code-sent {
+  min-height: 250px;
+  justify-content: space-evenly;
+}
+
+@media screen and (min-width: 600px) {
+  .action-container {
+    min-height: 125px;
+  }
+}
